refactor: Convert EditTsumegoTest to typed EditTestCase objects fix: Add type safety casts for Model::find('count') results fix: Correct Controller $params array access syntax fix: Fix RatingAchievementTest to properly test unlock conditions fix: Correct HeroPowersTest typo 'user-up' → 'used-up' fix: Use count() for array comparison in TsumegoXPAndRatingTest fix: Set CURLOPT_SSL_VERIFYHOST to proper int type in WebPServingTest refactor: Remove invalid @param PHPDoc tags refactor: Remove redundant assertions flagged by PHPStan docs: Update phpstan.neon comments for EditTestCase refactor

+++ src/Utility/TimeMode.php
@@ -39,7 +39,7 @@ class TimeMode
 			return;
 		}
 
-		$this->currentOrder = ClassRegistry::init('TimeModeAttempt')->find('count', [
+		$this->currentOrder = (int) ClassRegistry::init('TimeModeAttempt')->find('count', [
 			'conditions' => [
 				'time_mode_session_id' => $this->currentSession['TimeModeSession']['id'],
 				'time_mode_attempt_status_id !=' => TimeModeUtil::$ATTEMPT_RESULT_QUEUED]]) + 1;
